Extracted GA4GHBeaconResponse class
